requirements.txt
flask
flask-cors
ultralytics
pillow
opencv-python-headless
gunicorn
numpy